Fixed header to match function

+++ setup.h
@@ -1,1 +1,1 @@
-void setupSimulation(char * fileName, boidContainer * boids, int ** board);
+void setupSimulation(char * fileName, boidContainer * boids, short ** board);
